Move aws-cache tests to source code
