+++ package.json
@@ -76,6 +76,7 @@
     "@types/tar": "6.1.6",
     "@typescript-eslint/eslint-plugin": "5.18.0",
     "@typescript-eslint/parser": "5.18.0",
+    "aws-sdk-client-mock": "^3.0.0",
     "eslint": "8.13.0",
     "eslint-config-prettier": "8.5.0",
     "eslint-plugin-import": "2.25.4",
